Adjust the event listener on the plus btn
When the button is clicked push currentNumDisVal into the values array
and clear the value in the currentNumDis.
Check if there are any numbers in the fullOperationsDis and if there is
none, make sure no plus is added with the number being displayed in the
fullOperationsDis. If there is, add a plus right before displaying the
number.