Remove values array and add num1 and num2 vars
num1 and num2 will hold the first and second inputs that
the operation will be done on.
create a doOperation function that takes in two parameters.
the operation symbol and the actual operation.
The function first checks if num2 is undefined, and if so,
it checks if num1 exists, if it does, display num1
and the sign operation symbol in the fullOperationDis.
if num1 does not exist, set num1 to the currentNumDisVal
and then display num1 with the operation symbol after in the
fullOperationDis.

After this, set currentNumDis to nothing and assign the
operation variable to the op parameter. this will depend on each
operation.

Craete an event listner that runs the doOperation function
for each of the operations (plus,minus,divide, and times).

Create an event listener to the equals buttons. When clicked,
check if num1 exists, if it does, set num2 to currentNumDisVal
display num2 in the fullOperationDis. Then check if num2 is equals
to 0 and the operation is division, if so, give an error message.
if not create a result variable that will run the operate
function with the paremeters of the operation, num1 and num2.
set currentNumDis to the display the result. Set num1 to the
result, and num2 to undefined.